requirements.txt
absl-py==0.13.0
addict==2.4.0
aiohttp==3.7.4.post0
albumentations==1.0.2
alembic==1.6.5
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
bitsandbytes-cuda111==0.0.24
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cliff==3.8.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
dissect.cstruct==2.0
dominate==2.6.0
easydict==1.9
efficientnet-pytorch==0.7.1
feather-format==0.4.1
filelock==3.1.0
Flask==2.0.1
flatbuffers==1.12
fsspec==2021.9.0
future==0.18.2
gast==0.4.0
gevent==21.1.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
huggingface-hub==0.0.17
hydra-core==1.1.1
idna==3.2
imageio==2.9.0
imgutils==0.1.2
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.37.0
loguru==0.5.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
multidict==5.1.0
networkx==2.6.1
ninja==1.10.2
nnAudio==0.2.6
numpy==1.19.3
oauthlib==3.1.1
omegaconf==2.1.1
onnx==1.10.2
onnx-simplifier==0.3.5
onnxoptimizer==0.2.6
onnxruntime==1.8.0
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
optuna==2.9.1
packaging==21.0
pandas==1.3.0
pbr==5.6.0
Pillow==8.3.1
pooch==1.5.1
pretty-errors==1.2.24
prettytable==2.1.0
progressbar==2.5
progressbar2==3.53.1
protobuf==3.17.3
psutil==5.8.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.1
python-editor==1.0.4
python-utils==2.5.6
pytorch-lightning==1.5.0
pytorch-ranger==0.1.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rich==10.12.0
rsa==4.7.2
sacremoses==0.0.46
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
six==1.15.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.22
stevedore==3.3.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminaltables==3.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2021.7.2
timm==0.4.12
tokenizers==0.10.3
torch==1.9.0
torch-optimizer==0.1.0
torchaudio==0.9.0
torchfile==0.1.0
torchinfo==1.5.3
torchmetrics==0.5.1
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
transformers==4.11.0
typing-extensions==3.7.4.3
urllib3==1.26.6
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
zope.event==4.5.0
zope.interface==5.4.0
